backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.2.0,<3.0
numpy>=1.26.2
ta==0.10.2
aiohttp==3.9.1
python-multipart==0.0.6
sqlalchemy==2.0.23
python-dateutil==2.8.2
anthropic==0.21.0
apscheduler==3.10.4
pytz==2024.1
